Separate data for every user (#69)
* User will now be saved in the DB

* Home controller will now get the userID and gives thes to the Home Service. Those test will now alos accept another argument.

* Home endpoint now is correct. Gives the user access to creating and edeting a home

* Deletion of unused Inports and varibels

* Added Test for UserDoesNotHavePremision

* Added Test to Oauth

* Added UserID to Test and to Controller endpoints

* TaskMembers will be saved in the taskSeries isnted of task Definition. Might be usefull to store all assotiated with the task in a set. so i can add all home members and then add other usersers

* Adding a Task will now throw a Exeption if you are not allowed to. Also the user will be added as responisble to a task

* Deletion of unused Imports

* fixes to old methods in the test class

* Change to the TaskServiceTest

* Item has now seperate access Points with added security. Also deletion of unused Imports

* Update of frontend DTOs to match the backend

* Update so task will only appear ones. Use of a Set to delete dupplications

* By creating a Item you now have to chose a Home as well

* Creation of Endpoint to get all Item Names

* Now the user has the option to add a item to a task

+++ backend/src/main/java/org/example/backend/service/HomeService.java
@@ -11,6 +11,7 @@ import org.example.backend.repro.HomeRepro;
 import org.example.backend.service.mapper.HomeMapper;
 import org.example.backend.service.security.IdService;
 import org.example.backend.service.security.exception.HomeDoesNotExistException;
+import org.example.backend.service.security.exception.UserDoesNotHavePermissionException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,19 +30,21 @@ public class HomeService {
         this.idService = idService;
     }
 
-    public List<HomeTableReturnDTO> getAllHomes() {
-        return homeRepro.findAll().stream()
+    public List<HomeTableReturnDTO> getAllHomes(String userId) {
+        return homeRepro.findHomesByMemberId(userId).stream()
                 .map(homeMapper::mapToHomeTableReturn).toList();
-
     }
 
-    public HomeTableReturnDTO createNewHome(CreateHomeDTO createHomeDTO) {
+    public HomeTableReturnDTO createNewHome(String userId, CreateHomeDTO createHomeDTO) {
         Home home = homeMapper.mapToHome(createHomeDTO).withId(idService.createNewId());
+        home.members().put(userId, Role.ADMIN);
         return homeMapper.mapToHomeTableReturn(homeRepro.save(home));
     }
 
-    public HomeTableReturnDTO editHome(String id, EditHomeDTO editHomeDTO) throws HomeDoesNotExistException {
+    public HomeTableReturnDTO editHome(String userId,String id, EditHomeDTO editHomeDTO) throws HomeDoesNotExistException {
         Home home = homeRepro.findById(id).orElseThrow(() ->(new HomeDoesNotExistException("No Home with this ID found")));
+        //Chek if User has permission
+        checkForPermission(home,userId);
 
         //Changes Values if Change is provided by User
         if(editHomeDTO.name() != null){
@@ -52,12 +55,6 @@ public class HomeService {
             home = home.withAddress(editHomeDTO.address());
         }
 
-        if(editHomeDTO.associatedUsers() != null){
-            for(String userId : editHomeDTO.associatedUsers()){
-                home.members().put(userId, Role.MEMBER);
-            }
-        }
-
         //Saves Home
         homeRepro.save(home);
 
@@ -65,12 +62,16 @@ public class HomeService {
 
     }
 
-    public void deleteHome(String id) {
+    public void deleteHome(String userId,String id) {
+        //TODO What happens to connect Task
+        Home home = homeRepro.findById(id).orElseThrow(() ->(new HomeDoesNotExistException("No Home with this ID found")));
+        //Chek if User has permission
+        checkForPermission(home,userId);
         homeRepro.deleteById(id);
     }
 
-    public List<HomeListReturnDTO> getHomeNames() {
-        return homeRepro.findAll().stream()
+    public List<HomeListReturnDTO> getHomeNames(String userId) {
+        return homeRepro.findHomesByMemberId(userId).stream()
                 .map(homeMapper::mapToHomeListReturn)
                 .toList();
     }
@@ -78,5 +79,11 @@ public class HomeService {
 
     public List<String> findHomeConnectedToUser(String userId) {
         return homeRepro.findHomesByMemberId(userId).stream().map(Home::id).toList();
+    }
+
+    private void checkForPermission(Home home, String userId){
+        if(!home.members().containsKey(userId)){
+            throw new UserDoesNotHavePermissionException("User Does Not Have Permission");
+        }
     }
 }
